Refactor types and decouple CronExpression from CronExpressionParser to avoid any internal circular dependencies
