+++ src/index.ts
@@ -1,46 +1,11 @@
 /* istanbul ignore file */
+export * from './fields';
 export { CronFieldCollection } from './CronFieldCollection';
 export { CronParser } from './CronParser';
 export { CronDate } from './CronDate';
-export { CronDayOfMonth } from './fields/CronDayOfMonth';
-export { CronDayOfWeek } from './fields/CronDayOfWeek';
-export { CronField } from './fields/CronField';
-export { CronHour } from './fields/CronHour';
-export { CronMinute } from './fields/CronMinute';
-export { CronMonth } from './fields/CronMonth';
-export { CronSecond } from './fields/CronSecond';
-export { CronExpression } from './CronExpression';
+export { CronExpression, CronExpressionOptions } from './CronExpression';
 import { CronParser } from './CronParser';
 
 export default CronParser;
 
-// Exported types
-export {
-  // Enums
-  DateMathOp,
-  TimeUnit,
-  DayOfWeek,
-  PredefinedExpressions,
-
-  // Interfaces
-  CronFields,
-  CronOptions,
-  FieldRange,
-
-  // Types
-  CronChars,
-  CronConstraints,
-  CronFieldType,
-  CronMax,
-  CronMin,
-  DayOfMonthRange,
-  DayOfWeekRange,
-  HourRange,
-  IntRange,
-  MonthRange,
-  ParseStringResponse,
-  RangeFrom,
-  SerializedCronField,
-  SerializedCronFields,
-  SixtyRange,
-} from './types';
+export * from './fields';
